requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
psycopg==3.2.9
psycopg-binary==3.2.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
six==1.17.0
sqlparse==0.5.3
